Created various C# scripts
Created the baseline C# scripts that will be used later. Created Abstract classes for all types of cards, that are all subclasses of the foundational AbstractCrad. Created a static player class which will hold any data between encounters and the player's total deck. As well, created a Demo script that runs on startup and creates an Encounter instance which will be the "proof of concept" encounter to be finished by August.