Created Card Prefab and adjusted Encounter
Added a Prefab Game Object which holds the CardPrefab script. Whenever a new card is in the player hand, a new instance of the prefab will be created with the set data of whatever card is drawn. Also, now there is EncounterControl script which takes in an Encounter object to start a combat. It will load visuals and manage the combat flow based on the conditions of the passed instance of the Encounter class.