Scene Integration and flight paths
Added a function to transition between scenes when space bar is clicked. It loads both scenes at the same time and disables all game objects in the scene it is leaving. Also created a file to hold cross scene data. Also changed how bullets decide their next position. It now calls a method that is implemented by the bullet, so each bullet can move differently. Added an arch to the tomahawk bullet.